set up for deployment with heroku-cra-node; moved front-end code into this repo
