+++ package.json
@@ -3,13 +3,21 @@
   "version": "1.0.0",
   "description": "Backend server for the Riverbed application.",
   "main": "server.js",
+  "engines": {
+    "node": "6.9.x"
+  },
   "scripts": {
     "pretest": "npm run lint",
     "test": "mocha",
     "test:watch": "npm run test:run -- -w",
-    "start": "nodemon server.js",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "start": "node server",
+    "heroku-postbuild": "cd react-ui/ && npm install --only=dev && npm install && npm run build"
   },
+  "cacheDirectories": [
+    "node_modules",
+    "react-ui/node_modules"
+  ],
   "author": "",
   "license": "MIT",
   "dependencies": {
