+++ server/auth/ensure-auth.js
